services/python/requirements.txt
kafka-python==2.0.2
psycopg[binary]==3.1.19
python-dotenv==1.0.0
